# artifacts/api-server/src/routes/products.ts
import { Router, type IRouter } from "express";
import { eq, and, gte, lte, desc, asc, ilike, count, sql } from "drizzle-orm";
import { db, productsTable, categoriesTable, reviewsTable } from "@workspace/db";
import {
  ListProductsQueryParams,
  ListProductsResponse,
  CreateProductBody,
  GetProductParams,
  GetProductResponse,
  UpdateProductParams,
  UpdateProductBody,
  UpdateProductResponse,
  DeleteProductParams,
  GetFeaturedProductsResponse,
  GetPopularProductsResponse,
  GetNewArrivalsResponse,
  GetRelatedProductsParams,
  GetRelatedProductsResponse,
  GetPopularBrandsResponse,
} from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/products/featured", async (_req, res): Promise<void> => {
  const rows = await db
    .select()
    .from(productsTable)
    .where(eq(productsTable.isFeatured, true))
    .orderBy(desc(productsTable.rating))
    .limit(8);
  const withCategory = rows.map((r) => ({ ...r, categoryName: null }));
  res.json(GetFeaturedProductsResponse.parse(withCategory));
});

router.get("/products/popular", async (_req, res): Promise<void> => {
  const rows = await db
    .select()
    .from(productsTable)
    .orderBy(desc(productsTable.reviewCount), desc(productsTable.rating))
    .limit(8);
  const withCategory = rows.map((r) => ({ ...r, categoryName: null }));
  res.json(GetPopularProductsResponse.parse(withCategory));
});

router.get("/products/new-arrivals", async (_req, res): Promise<void> => {
  const rows = await db
    .select()
    .from(productsTable)
    .where(eq(productsTable.isNew, true))
    .orderBy(desc(productsTable.createdAt))
    .limit(8);
  const withCategory = rows.map((r) => ({ ...r, categoryName: null }));
  res.json(GetNewArrivalsResponse.parse(withCategory));
});

router.get("/products", async (req, res): Promise<void> => {
  const qp = ListProductsQueryParams.safeParse(req.query);
  if (!qp.success) {
    res.status(400).json({ error: qp.error.message });
    return;
  }

  const { categoryId, search, minPrice, maxPrice, brand, inStock, sortBy, page = 1, limit = 12 } = qp.data;

  const conditions = [];
  if (categoryId) conditions.push(eq(productsTable.categoryId, categoryId));
  if (search) conditions.push(ilike(productsTable.name, `%${search}%`));
  if (minPrice != null) conditions.push(gte(productsTable.price, minPrice));
  if (maxPrice != null) conditions.push(lte(productsTable.price, maxPrice));
  if (brand) conditions.push(ilike(productsTable.brand, `%${brand}%`));
  if (inStock != null) conditions.push(eq(productsTable.inStock, inStock));

  const whereClause = conditions.length > 0 ? and(...conditions) : undefined;

  let orderByClause;
  switch (sortBy) {
    case "price_asc":
      orderByClause = asc(productsTable.price);
      break;
    case "price_desc":
      orderByClause = desc(productsTable.price);
      break;
    case "rating":
      orderByClause = desc(productsTable.rating);
      break;
    case "newest":
      orderByClause = desc(productsTable.createdAt);
      break;
    default:
      orderByClause = desc(productsTable.id);
  }

  const offset = (Number(page) - 1) * Number(limit);

  const [rows, totalCount] = await Promise.all([
    db.select().from(productsTable).where(whereClause).orderBy(orderByClause).limit(Number(limit)).offset(offset),
    db.select({ count: count() }).from(productsTable).where(whereClause),
  ]);

  const withCategory = rows.map((r) => ({ ...r, categoryName: null }));

  res.json(
    ListProductsResponse.parse({
      items: withCategory,
      total: totalCount[0]?.count ?? 0,
      page: Number(page),
      limit: Number(limit),
    }),
  );
});

router.post("/products", async (req, res): Promise<void> => {
  const parsed = CreateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [row] = await db.insert(productsTable).values(parsed.data).returning();
  res.status(201).json({ ...row, categoryName: null });
});

router.get("/products/:id", async (req, res): Promise<void> => {
  const params = GetProductParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [row] = await db.select().from(productsTable).where(eq(productsTable.id, params.data.id));
  if (!row) {
    res.status(404).json({ error: "Product not found" });
    return;
  }

  const related = await db
    .select()
    .from(productsTable)
    .where(and(eq(productsTable.categoryId, row.categoryId), sql`${productsTable.id} != ${row.id}`))
    .limit(4);

  res.json(
    GetProductResponse.parse({
      ...row,
      categoryName: null,
      relatedProducts: related.map((r) => ({ ...r, categoryName: null })),
    }),
  );
});

router.patch("/products/:id", async (req, res): Promise<void> => {
  const params = UpdateProductParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [row] = await db
    .update(productsTable)
    .set(parsed.data)
    .where(eq(productsTable.id, params.data.id))
    .returning();
  if (!row) {
    res.status(404).json({ error: "Product not found" });
    return;
  }
  res.json(UpdateProductResponse.parse({ ...row, categoryName: null }));
});

router.delete("/products/:id", async (req, res): Promise<void> => {
  const params = DeleteProductParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  await db.delete(productsTable).where(eq(productsTable.id, params.data.id));
  res.sendStatus(204);
});

router.get("/products/:id/related", async (req, res): Promise<void> => {
  const params = GetRelatedProductsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [product] = await db.select().from(productsTable).where(eq(productsTable.id, params.data.id));
  if (!product) {
    res.json(GetRelatedProductsResponse.parse([]));
    return;
  }
  const rows = await db
    .select()
    .from(productsTable)
    .where(and(eq(productsTable.categoryId, product.categoryId), sql`${productsTable.id} != ${product.id}`))
    .limit(6);
  res.json(GetRelatedProductsResponse.parse(rows.map((r) => ({ ...r, categoryName: null }))));
});

router.get("/stats/brands", async (_req, res): Promise<void> => {
  const rows = await db
    .select({ brand: productsTable.brand, productCount: count() })
    .from(productsTable)
    .groupBy(productsTable.brand)
    .orderBy(desc(count()))
    .limit(10);
  res.json(GetPopularBrandsResponse.parse(rows.map((r) => ({ ...r, logoUrl: null }))));
});

export default router;