# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Velik e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: brand
          in: query
          schema:
            type: ["string", "null"]
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
            enum: ["price_asc", "price_desc", "rating", "newest", null]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ─── Reviews ──────────────────────────────────────────────────────────────
  /products/{productId}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── Cart ─────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin) or current user orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── Auth ─────────────────────────────────────────────────────────────────
...
  /repair-requests:
    get:
      operationId: listRepairRequests
      tags: [repair]
      summary: List repair requests
      responses:
        "200":
          description: Repair requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepairRequest"
    post:
      operationId: createRepairRequest
      tags: [repair]
      summary: Submit repair request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairRequest"
  /repair-requests/{id}:
    patch:
      operationId: updateRepairRequest
      tags: [admin]
      summary: Update repair request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairRequest"

  # ─── Promotions ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, categoryId, brand, inStock, imageUrl, rating, reviewCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        brand:
          type: string
        inStock:
          type: boolean
        stockCount:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        badge:
          type: ["string", "null"]
          description: "ХИТ ПРОДАЖ | ТОП ВЫБОР | НОВИНКА | ЛУЧШАЯ ЦЕНА | ЭКО"
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        discountPercent:
          type: ["integer", "null"]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            specs:
              type: object
              additionalProperties:
                type: string
            relatedProducts:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, slug, price, categoryId, brand, imageUrl]
      properties:
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        brand:
          type: string
        inStock:
          type: boolean
        stockCount:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        specs:
          type: object
          additionalProperties:
            type: string
        badge:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isNew:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        brand:
          type: string
        inStock:
          type: boolean
        stockCount:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        specs:
          type: object
          additionalProperties:
            type: string
        badge:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isNew:
          type: boolean
    Review:
      type: object
      required: [id, productId, authorName, rating, text, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        text:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [authorName, rating, text]
      properties:
        authorName:
          type: string
        rating:
          type: integer
        text:
          type: string
...
    Order:
      type: object
      required: [id, status, total, createdAt, items]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        total:
          type: number
        createdAt:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        deliveryMethod:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
          enum: [online, cash]
        comment:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, productId, quantity, price, productName, productImageUrl]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        productName:
          type: string
        productImageUrl:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerPhone, deliveryMethod, paymentMethod, items]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        deliveryMethod:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
          enum: [online, cash]
        comment:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity, price]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
...
    RepairRequest:
      type: object
      required: [id, customerName, customerPhone, bikeDescription, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        bikeDescription:
          type: string
        problemDescription:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, in_progress, completed, cancelled]
        estimatedCost:
          type: ["number", "null"]
        createdAt:
          type: string
    RepairRequestInput:
      type: object
      required: [customerName, customerPhone, bikeDescription]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        bikeDescription:
          type: string
        problemDescription:
          type: ["string", "null"]
    RepairRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, in_progress, completed, cancelled]
        estimatedCost:
          type: ["number", "null"]
...