# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Velik e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDeliveryMethod } from './orderDeliveryMethod';
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  total: number;
  createdAt: string;
  customerName?: string;
  customerPhone?: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  deliveryMethod?: OrderDeliveryMethod;
  paymentMethod?: OrderPaymentMethod;
  /** @nullable */
  comment?: string | null;
  items: OrderItem[];
}