# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Velik e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateSpecs } from './productUpdateSpecs';

export interface ProductUpdate {
  name?: string;
  price?: number;
  /** @nullable */
  oldPrice?: number | null;
  categoryId?: number;
  brand?: string;
  inStock?: boolean;
  stockCount?: number;
  imageUrl?: string;
  images?: string[];
  /** @nullable */
  description?: string | null;
  specs?: ProductUpdateSpecs;
  /** @nullable */
  badge?: string | null;
  isFeatured?: boolean;
  isNew?: boolean;
}