# lib/api-zod/src/generated/types/repairRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Velik e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface RepairRequestInput {
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  bikeDescription: string;
  /** @nullable */
  problemDescription?: string | null;
}